requirements.txt
fastapi
uvicorn
requests
pillow
google-genai
python-multipart
holidays